+++ R/register.R
@@ -34,7 +34,7 @@
 #'
 #' # cleanup
 #' unlink(dir, recursive = TRUE)
-cpp_register <- function(path = ".", quiet = FALSE) {
+cpp_register <- function(path = ".", quiet = !is_interactive()) {
   stop_unless_installed(get_cpp_register_needs())
 
   r_path <- file.path(path, "R", "cpp11.R")
@@ -138,7 +138,7 @@ cpp_register <- function(path = ".", quiet = FALSE) {
 
 utils::globalVariables(c("name", "return_type", "line", "decoration", "context", ".", "functions", "res"))
 
-get_registered_functions <- function(decorations, tag, quiet = FALSE) {
+get_registered_functions <- function(decorations, tag, quiet = !is_interactive()) {
   if (NROW(decorations) == 0) {
     return(tibble::tibble(file = character(), line = integer(), decoration = character(), params = list(), context = list(), name = character(), return_type = character(), args = list()))
   }